requirements.txt
tk
pygame 
ctypes-callable
comtypes
pycaw
threaded
Pillow
pystray